+++ tests/Unit/Public/Assert-ElevatedUser.Tests.ps1
@@ -24,7 +24,7 @@ BeforeDiscovery {
 }
 
 BeforeAll {
-    $script:dscModuleName = 'DSCResource.Common'
+    $script:dscModuleName = 'DscResource.Common'
 
     Import-Module -Name $script:dscModuleName
 
@@ -61,17 +61,15 @@ Describe 'Assert-ElevatedUser' -Tag 'Public' {
     }
 
     It 'Should throw the correct error' -Skip:$mockIsElevated {
-        InModuleScope -ScriptBlock {
-            $mockErrorMessage = $script:localizedData.ElevatedUser_UserNotElevated
-
-            { Assert-ElevatedUser } | Should -Throw -ExpectedMessage $mockErrorMessage
+        $mockErrorMessage = InModuleScope -ScriptBlock {
+            $script:localizedData.ElevatedUser_UserNotElevated
         }
+
+        { Assert-ElevatedUser } | Should -Throw -ExpectedMessage $mockErrorMessage
     }
 
     It 'Should not throw an exception' -Skip:(-not $mockIsElevated) {
-        InModuleScope -ScriptBlock {
-            { Assert-ElevatedUser } | Should -Not -Throw
-        }
+        { Assert-ElevatedUser } | Should -Not -Throw
     }
 
     Context 'When on Linux or macOS' {
@@ -106,9 +104,7 @@ Describe 'Assert-ElevatedUser' -Tag 'Public' {
         }
 
         It 'Should not throw an exception' {
-            InModuleScope -ScriptBlock {
-                { Assert-ElevatedUser } | Should -Not -Throw
-            }
+            { Assert-ElevatedUser } | Should -Not -Throw
         }
     }
 }